add resolve handler (no persistence)
Add the /resolve/#String route that takes a URL and resolves
redirects.  Return 200 text/plain with the final URL (which may be
the input), 404 if the given URL is valid but cannot be resolved to
a final URL for some reason (reason included in response body), or
400 if the URL cannot be parsed.

License shortcircuit-server under AGPL3.